+++ src/components/organisms/GiftRankItems/index.tsx
@@ -1,5 +1,6 @@
 import styled from '@emotion/styled';
 import { useState } from 'react';
+import { useLoaderData } from 'react-router-dom';
 
 import { Button } from '@/components/atoms/Button';
 import type { RankingGoodsItemsProps } from '@/components/molecules/GoodsItem/Ranking';
@@ -7,23 +8,9 @@ import { RankingGoodsItems } from '@/components/molecules/GoodsItem/Ranking';
 import { GridItemsBox } from '@/components/molecules/GridItemsBox';
 import type { IteratingItemProp } from '@/components/molecules/types/IteratingItemProp';
 
-const item = {
-  name: 'BBQ 양념치킨+크림치즈볼+콜라1.25L',
-  imageSrc:
-    'https://st.kakaocdn.net/product/gift/product/20231030175450_53e90ee9708f45ffa45b3f7b4bc01c7c.jpg',
-  subtitle: 'BBQ',
-  title: `BBQ 양념치킨+크림치즈볼+콜라1.25L`,
-  amount: 29000,
-};
-
-const mockData: (RankingGoodsItemsProps & IteratingItemProp)[] = [];
-
-for (let i = 0; i < 21; i++) {
-  mockData.push({ rankingIndex: i + 1, ...item, id: `${i}` });
-}
-
 export const GiftRankItems = () => {
   const [isClicked, setIsClicked] = useState<boolean>(false);
+  const mockData = useLoaderData() as (RankingGoodsItemsProps & IteratingItemProp)[];
   const showData = isClicked ? mockData : mockData.slice(0, 6);
 
   return (
